refactor(state): extract helpers (WP5.9.6.6c)
- Create useRegistrationHelpers module for helper functions
- Move getCourtData, markUserTyping, clearSuccessResetTimer
- Move getCourtsOccupiedForClearing, guardAddPlayerEarly, guardAgainstGroupDuplicate
- Move validateGroupCompat (exported standalone)
- Update useRegistrationAppState to call useRegistrationHelpers
- Add exports to barrel file

Main file: 541 → 398 lines (-143 lines, -26%)
Helpers module: 192 lines

TARGET ACHIEVED: <500 lines (398)
