+++ src/registration/appHandlers/state/index.js
@@ -3,4 +3,5 @@ export { useRegistrationRuntime } from './useRegistrationRuntime.js';
 export { useRegistrationDataLayer } from './useRegistrationDataLayer.js';
 export { useRegistrationDomainHooks } from './useRegistrationDomainHooks.js';
 export { useRegistrationDerived } from './useRegistrationDerived.js';
+export { useRegistrationHelpers, validateGroupCompat } from './useRegistrationHelpers.js';
 export { buildRegistrationReturn } from './buildRegistrationReturn.js';
